+++ tests/Pest.php
@@ -1,5 +1,7 @@
 <?php
 
 use Tests\TestCase;
+use Tests\UnitTestCase;
 
-pest()->extends(TestCase::class)->in('Unit', 'Feature');
+pest()->extends(TestCase::class)->in('Feature');
+pest()->extends(UnitTestCase::class)->in('Unit');