+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [1.1.3] - 2025-11-
+- Save integers for settings enums
 
 ## [1.1.2] - 2025-11-09
 - Added extension hook to Ui5Registry