+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [1.1.3] - 2025-11-
 - Save enums as objects in registry
 - Renamed visibilityRole to role in Setting
+- Enhanced docs for Role attribute
 
 ## [1.1.2] - 2025-11-09
 - Added extension hook to Ui5Registry