+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Save enums as objects in registry
 - Renamed visibilityRole to role in Setting
 - Enhanced Role attribute
+- Added note property to Setting attribute
 
 ## [1.1.2] - 2025-11-09
 - Added extension hook to Ui5Registry