+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.1.3] - 2025-11-
+## [1.1.4] - 2025-11-
+
+## [1.1.3] - 2025-11-11
 - Save enums as objects in registry
 - Renamed visibilityRole to role in Setting
 - Enhanced Role attribute